refactor(ui): Разложил Stepper по компонентам, тайпинги
