+++ packages/ui/src/components/Stepper/Stepper.mdx
@@ -7,9 +7,10 @@ route: /components/Stepper
 import { Playground, Props } from 'docz'
 import { IconMinus, IconPlus } from '@sberdevices/plasma-icons'
 
-import { Stepper, StepperButton, StepperValue, StepperRoot } from './Stepper'
 import { ThemeSelect } from '../../helpers/ThemeSelect'
 import { ThemeBackground } from '../../helpers/ThemeBackground'
+
+import { Stepper, StepperButton, StepperValue, StepperRoot } from '.'
 
 <ThemeSelect />
