+++ packages/ui/src/components/Stepper/Stepper.tsx
@@ -1,98 +1,17 @@
-import React, { useCallback } from 'react';
-import styled, { css } from 'styled-components';
-import { typography, colors, scalingPixelBasis } from '@sberdevices/plasma-tokens';
+import React from 'react';
 import { IconMinus, IconPlus, IconTrash } from '@sberdevices/plasma-icons';
 
-import { PickOptional } from '../../types/PickOptional';
-import { ActionButton, ButtonProps } from '../Button';
+import { PickOptional } from '../../types';
 
-const StyledRoot = styled.div`
-    display: flex;
-    align-items: center;
-    box-sizing: border-box;
-    justify-content: space-between;
-`;
+import { StepperButton, StepperButtonProps, StepperRoot, StepperValue } from '.';
 
-interface StyledValueProps {
-    /**
-     * Неактивное состояние: состояние, при котором компонент отображается, но недоступен для действий пользователя
-     */
-    disabled?: boolean;
-    /**
-     * Состояние, когда значение контрола близко к предельному
-     */
-    isWarning?: boolean;
-}
-
-const StyledValue = styled.span<StyledValueProps>`
-    ${typography.body2}
-
-    box-sizing: border-box;
-
-    margin-left: ${12 / scalingPixelBasis}rem;
-    margin-right: ${12 / scalingPixelBasis}rem;
-    min-width: ${20 / scalingPixelBasis}rem;
-
-    color: ${colors.text};
-
-    text-align: center;
-
-    ${({ isWarning }) =>
-        isWarning &&
-        css`
-            color: ${colors.warning};
-        `}
-
-    ${({ disabled }) =>
-        disabled &&
-        css`
-            opacity: 0.4;
-        `}
-`;
-
-export interface StepperButtonProps
-    extends PickOptional<
-        ButtonProps,
-        'pin' | 'view' | 'disabled' | 'className' | 'style' | 'onFocus' | 'onBlur' | 'onClick'
-    > {
-    icon?: React.ReactElement;
-}
-
-export interface StepperValueProps extends Pick<StyledValueProps, 'disabled' | 'isWarning'> {
-    /**
-     * Выводимое значение
-     */
-    value: number;
-}
-
-export const StepperRoot = StyledRoot;
-
-export const StepperButton: React.FC<StepperButtonProps> = ({
-    pin = 'circle-circle',
-    view = 'secondary',
-    icon,
-    ...rest
-}) => (
-    <ActionButton size="m" pin={pin} view={view} {...rest}>
-        {icon}
-    </ActionButton>
-);
-
-export const StepperValue: React.FC<StepperValueProps> = ({ value, disabled, isWarning }) => (
-    <StyledValue disabled={disabled} isWarning={isWarning}>
-        {value}
-    </StyledValue>
-);
-
-export interface StepperProps {
+export interface StepperProps
+    extends Omit<React.HTMLAttributes<HTMLDivElement>, 'onChange' | 'onFocus' | 'onBlur'>,
+        PickOptional<StepperButtonProps, 'onFocus' | 'onBlur'> {
     /**
      * Числовое значение
      */
     value: number;
-    /**
-     * Обработчик изменения значения счетчика
-     */
-    onChange: (value: number) => void;
     /**
      * Шаг изменения значения
      */
@@ -113,15 +32,14 @@ export interface StepperProps {
      * При достижении минимального количества, кнопка минус превратится в удалить
      */
     remover?: boolean;
+    /**
+     * Обработчик клика по кнопки удаления
+     */
     onRemove?: React.MouseEventHandler<HTMLButtonElement>;
     /**
-     * Обработчик фокуса по кнопкам +/-
+     * Обработчик изменения значения счетчика
      */
-    onFocus?: React.FocusEventHandler<HTMLButtonElement>;
-    /**
-     * Обработчик ухода фокуса от кнопок +/-
-     */
-    onBlur?: React.FocusEventHandler<HTMLButtonElement>;
+    onChange: (value: number) => void;
 }
 
 export const Stepper: React.FC<StepperProps> = ({
@@ -135,17 +53,18 @@ export const Stepper: React.FC<StepperProps> = ({
     onFocus,
     onBlur,
     onRemove,
+    ...rest
 }) => {
-    const onLessClick = useCallback(() => onChange(value - step), [value, step, onChange]);
-    const onMoreClick = useCallback(() => onChange(value + step), [value, step, onChange]);
-    const onRemoveClick = useCallback((e) => onRemove?.(e), [onRemove]);
+    const onLessClick = React.useCallback(() => onChange(value - step), [value, step, onChange]);
+    const onMoreClick = React.useCallback(() => onChange(value + step), [value, step, onChange]);
+    const onRemoveClick = React.useCallback((e) => onRemove?.(e), [onRemove]);
     const isMin = value <= min;
     const isMax = value >= max;
     const lessDisabled = isMin || value - step < min;
     const moreDisabled = isMax || value + step > max;
 
     return (
-        <StepperRoot>
+        <StepperRoot {...rest}>
             <StepperButton
                 disabled={disabled || (!remover && lessDisabled)}
                 icon={isMin && remover ? <IconTrash /> : <IconMinus />}
