+++ packages/ui/src/components/Stepper/StepperButton.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import { PickOptional } from '../../types/PickOptional';
+import { ActionButton, ButtonProps } from '../Button';
+
+export interface StepperButtonProps
+    extends PickOptional<
+        ButtonProps,
+        'pin' | 'view' | 'disabled' | 'className' | 'style' | 'onFocus' | 'onBlur' | 'onClick'
+    > {
+    icon?: React.ReactElement;
+}
+
+export const StepperButton: React.FC<StepperButtonProps> = ({
+    pin = 'circle-circle',
+    view = 'secondary',
+    icon,
+    ...rest
+}) => (
+    <ActionButton size="m" pin={pin} view={view} {...rest}>
+        {icon}
+    </ActionButton>
+);
