+++ packages/ui/src/components/Stepper/StepperRoot.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const StepperRoot = styled.div`
+    display: flex;
+    align-items: center;
+    box-sizing: border-box;
+    justify-content: space-between;
+    width: max-content;
+`;
