+++ packages/ui/src/components/Stepper/index.ts
@@ -1,2 +1,10 @@
-export { Stepper, StepperButton, StepperRoot, StepperValue } from './Stepper';
-export type { StepperButtonProps, StepperProps, StepperValueProps } from './Stepper';
+export { Stepper } from './Stepper';
+export type { StepperProps } from './Stepper';
+
+export { StepperButton } from './StepperButton';
+export type { StepperButtonProps } from './StepperButton';
+
+export { StepperRoot } from './StepperRoot';
+
+export { StepperValue } from './StepperValue';
+export type { StepperValueProps } from './StepperValue';
